requirements.txt
text
flask
flask-cors
openpyxl
pillow